Fix leaked electron impl in types library
