+++ src/main_types.ts
@@ -1,8 +1,3 @@
 export * from './constants';
-export * from './models/DownloadManager';
+export type { DownloadManager, Download } from './models/DownloadManager';
 export type { ElectronAPI } from './preload';
-
-import { ElectronAPI } from './preload';
-declare global {
-  const electronAPI: ElectronAPI;
-}
